snapshot 23-03-06 10:13

+++ src/main/java/nl/ultimateapps/demoDrop/Services/UserService.java
@@ -2,6 +2,7 @@ package nl.ultimateapps.demoDrop.Services;
 
 import nl.ultimateapps.demoDrop.Dtos.input.UserInputDto;
 import nl.ultimateapps.demoDrop.Dtos.output.UserDto;
+import nl.ultimateapps.demoDrop.Dtos.output.UserPublicDto;
 import nl.ultimateapps.demoDrop.Exceptions.RecordNotFoundException;
 import nl.ultimateapps.demoDrop.Exceptions.UsernameNotFoundException;
 import nl.ultimateapps.demoDrop.Helpers.mappers.UserMapper;
@@ -11,9 +12,9 @@ import nl.ultimateapps.demoDrop.Repositories.UserRepository;
 import nl.ultimateapps.demoDrop.Utils.AuthHelper;
 import nl.ultimateapps.demoDrop.Utils.RandomStringGenerator;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-
 import java.time.Instant;
 import java.util.*;
 
@@ -32,34 +33,34 @@ public class UserService {
     @Setter
     private UserRepository userRepository;
 
-    public ArrayList<UserDto> getUsers(int limit) {
+    public ArrayList<UserPublicDto> getUserPublicDtos (int limit) {
+        ArrayList<User> userList = GetUsersFromRepo(limit);
+        ArrayList<UserPublicDto> userPublicDtoList = new ArrayList<>();
+        for (User user : userList) {
+            UserPublicDto userPublicDto = UserMapper.mapToPublicDto(user);
+            userPublicDtoList.add(userPublicDto);
+        }
+        return userPublicDtoList;
+    }
+
+    public UserPublicDto getUserPublicDto(String username) {
+        User user = getUserFromRepo(username);
+        return UserMapper.mapToPublicDto(user);
+    }
+
+    public ArrayList<UserDto> getUserDtos(int limit) {
+        ArrayList<User> userList = GetUsersFromRepo(limit);
         ArrayList<UserDto> userDtoList = new ArrayList<>();
-        Iterable<User> userIterable = userRepository.findAllByOrderByCreatedDateDesc();
-        ArrayList<User> userArrayList = new ArrayList<>();
-        userIterable.forEach(userArrayList::add);
-        int numResults = userArrayList.size();
-        if (limit == 0) {
-            // return full list
-            for (User user : userArrayList) {
-                UserDto userDto = UserMapper.mapToDto(user);
-                userDtoList.add(userDto);
-            }
-        } else {
-            // return limited list
-            for (int i = 0; i < (Math.min(numResults, limit)); i++) {
-                UserDto userDto = UserMapper.mapToDto(userArrayList.get(i));
-                userDtoList.add(userDto);
-            }
+        for (User user : userList) {
+            UserDto userDto = UserMapper.mapToDto(user);
+            userDtoList.add(userDto);
         }
         return userDtoList;
     }
 
-    public UserDto getUser(String username) {
-        Optional<User> user = userRepository.findById(username);
-        if (!user.isPresent()) {
-            return null;
-        }
-        return UserMapper.mapToDto(user.get());
+    public UserDto getUserDto(String username) {
+        User user = getUserFromRepo(username);
+        return UserMapper.mapToDto(user);
     }
 
     public boolean userExists(String username) {
@@ -79,6 +80,9 @@ public class UserService {
 
 
     public String updateUser(String username, UserDto userDto) {
+        if (!AuthHelper.checkAuthorization(username)) { // check associative authorization
+            throw new AccessDeniedException("User has insufficient rights to alter user details for user " + username);
+        }
         if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
         userDto.setPassword(passwordEncoder.encode(userDto.getPassword()));
         User user = userRepository.findById(username).get();
@@ -88,6 +92,9 @@ public class UserService {
     }
 
     public String changePassword(String username, UserInputDto userInputDto) {
+        if (!AuthHelper.checkAuthorization(username)) { // check associative authorization
+            throw new AccessDeniedException("User" + AuthHelper.getPrincipalUsername() + " has insufficient rights to change the password for user " + username);
+        }
         if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
         userInputDto.setPassword(passwordEncoder.encode(userInputDto.getPassword()));
         User user = userRepository.findById(username).get();
@@ -97,6 +104,9 @@ public class UserService {
     }
 
     public String changeEmail(String username, UserInputDto userInputDto) {
+        if (!AuthHelper.checkAuthorization(username)) { // check associative authorization
+            throw new AccessDeniedException("User "+ AuthHelper.getPrincipalUsername() +" has insufficient rights to change the email address for user " + username);
+        }
         if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
         User user = userRepository.findById(username).get();
         user.setEmail(userInputDto.getEmail());
@@ -105,6 +115,10 @@ public class UserService {
     }
 
     public boolean deleteUser(String username) {
+        if (!AuthHelper.checkAuthorization(username)) { // check associative authorization
+            throw new AccessDeniedException("User "+ AuthHelper.getPrincipalUsername() +" has insufficient rights to delete user " + username + " from the system");
+        }
+
         if (AuthHelper.checkAuthorization(username)) {
             userRepository.deleteById(username);
             return true;
@@ -112,6 +126,9 @@ public class UserService {
     }
 
     public Set<Authority> getAuthorities(String username) {
+        if (!AuthHelper.checkAuthorization(username)) { // check associative authorization
+            throw new AccessDeniedException("User "+ AuthHelper.getPrincipalUsername() +" has insufficient rights to retrieve authorities for user " + username);
+        }
         if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
         User user = userRepository.findById(username).get();
         UserDto userDto = UserMapper.mapToDto(user);
@@ -119,7 +136,6 @@ public class UserService {
     }
 
     public void addAuthority(String username, String authority) {
-
         if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
         User user = userRepository.findById(username).get();
         user.addAuthority(new Authority(username, authority));
@@ -133,16 +149,6 @@ public class UserService {
         user.removeAuthority(authorityToRemove);
         userRepository.save(user);
     }
-
-//    public User toUser(UserDto userDto) { // niet alle velden zijn meegenomen . Allenede velden die je meerstuurt bij het creeren vn ene gebruiker.
-//
-//        var user = new User();
-//        user.setUsername(userDto.getUsername());
-//        user.setPassword(userDto.getPassword());
-//        user.setEmail(userDto.getEmail());
-//        user.setCreatedDate(userDto.getCreatedDate());
-//        return user;
-//    }
 
     public User updateExistingUser(User user, UserDto userDto) {
         if (userDto.getUsername() != null) {
@@ -195,6 +201,33 @@ public class UserService {
         user.setEnabled(desiredStatus);
         userRepository.save(user);
         return checkAccountStatus(username);
+    }
+
+    private ArrayList<User> GetUsersFromRepo(int limit) {
+        Iterable<User> userIterable = userRepository.findAllByOrderByCreatedDateDesc();
+        ArrayList<User> fullUserList = new ArrayList<>();
+        userIterable.forEach(fullUserList::add);
+        int numResults = fullUserList.size();
+        if (limit == 0) {
+            // return full list
+            return fullUserList;
+        } else {
+            // return limited list
+            ArrayList<User> limitedUserList = new ArrayList<>();
+            for (int i = 0; i < (Math.min(numResults, limit)); i++) {
+                User user = fullUserList.get(i);
+                limitedUserList.add(user);
+            }
+            return limitedUserList;
+        }
+    }
+
+    // Private functions:
+    private User getUserFromRepo(String username) {
+        if (userRepository.findById(username).isEmpty()) {
+            throw new UsernameNotFoundException(username);
+        }
+        return userRepository.findById(username).get();
     }
 
 }
